Refactoring: create custom water law regulation file
