+++ envergo/moulinette/regulations/__init__.py
@@ -1,0 +1,125 @@
+import json
+from functools import cached_property
+
+from envergo.evaluations.models import RESULTS
+from envergo.geodata.utils import to_geojson
+
+
+class MoulinetteRegulation:
+    """Run the moulinette for a single regulation (e.g Loi sur l'eau)."""
+
+    criterion_classes = []
+
+    def __init__(self, data_catalog):
+        self.catalog = data_catalog
+        self.catalog.update(self.get_catalog_data())
+        self.criterions = [
+            Criterion(self.catalog) for Criterion in self.criterion_classes
+        ]
+
+    def get_catalog_data(self):
+        return {}
+
+    @cached_property
+    def result(self):
+        """Compute global result from individual criterions."""
+
+        results = [criterion.result for criterion in self.criterions]
+
+        if RESULTS.soumis in results:
+            result = RESULTS.soumis
+        elif RESULTS.action_requise in results:
+            result = RESULTS.action_requise
+        else:
+            result = RESULTS.non_soumis
+
+        return result
+
+    def __getattr__(self, attr):
+        """Returs the corresponding criterion.
+
+        Allows to do something like this:
+        moulinette.loi_sur_leau.zones_inondables to fetch the correct criterion.
+        """
+        return self.get_criterion(attr)
+
+    def get_criterion(self, criterion_slug):
+        """Return the regulation with the given slug."""
+
+        def select_criterion(criterion):
+            return criterion.slug == criterion_slug
+
+        criterion = next(filter(select_criterion, self.criterions), None)
+        return criterion
+
+
+class CriterionMap:
+    """Data for a map that will be displayed with Leaflet."""
+
+    def __init__(self, center, polygons, caption, sources):
+        self.center = center
+        self.polygons = polygons
+        self.caption = caption
+        self.sources = sources
+
+    def to_json(self):
+
+        # Don't display full polygons
+        EPSG_WGS84 = 4326
+        buffer = self.center.buffer(500).transform(EPSG_WGS84, clone=True)
+
+        data = json.dumps(
+            {
+                "center": to_geojson(self.center),
+                "polygons": [
+                    {
+                        "polygon": to_geojson(polygon["polygon"].intersection(buffer)),
+                        "color": polygon["color"],
+                        "label": polygon["label"],
+                    }
+                    for polygon in self.polygons
+                ],
+                "caption": self.caption,
+                "sources": [
+                    {"name": map.name, "url": map.source} for map in self.sources
+                ],
+            }
+        )
+        return data
+
+
+class MoulinetteCriterion:
+    """Run a single moulinette check."""
+
+    def __init__(self, data_catalog):
+        self.catalog = data_catalog
+        self.catalog.update(self.get_catalog_data())
+
+    def get_catalog_data(self):
+        return {}
+
+    @cached_property
+    def result(self):
+        raise NotImplementedError("Implement the `result` method in the subclass.")
+
+    @property
+    def result_code(self):
+        """Return a unique code for the criterion result.
+
+        Sometimes, a same criterion can have the same result for different reasons.
+        Because of this, we want unique codes to display custom messages to
+        the user.
+        """
+
+        return self.result
+
+    @cached_property
+    def map(self):
+        try:
+            map = self._get_map()
+        except:  # noqa
+            map = None
+        return map
+
+    def _get_map(self):
+        return None
